requirements.txt
django>=1.3
git+https://github.com/BayCitizen/django-embedly.git
git+https://github.com/sunlightlabs/django-secretballot.git
git+https://github.com/bradjasper/django-jsonfield.git
django-tagging
django-tagging-autocomplete
django-form-utils
svn+http://django-tagging-autocomplete.googlecode.com/svn/trunk/